rename png -> png_file
